+++ aws/spec/images_query.go
@@ -36,6 +36,8 @@ import (
 //
 // - redhat::::instance-store
 //
+// - centos:centos
+//
 // The default values are: Arch="x86_64", Virt="hvm", Store="ebs"
 type ImageResolver func(*ec2.DescribeImagesInput) (*ec2.DescribeImagesOutput, error)
 
@@ -212,6 +214,7 @@ var (
 		"redhat":      RedHat,
 		"debian":      Debian,
 		"amazonlinux": AmazonLinux,
+		"centos":      CentOS,
 		"suselinux":   SuseLinux,
 		"windows":     Windows,
 	}
@@ -234,6 +237,13 @@ var (
 		Name: "debian", Id: "379101102735", DistroName: "debian", LatestVariant: "stretch",
 		MatchFunc: func(s string, d Distro) bool {
 			return strings.HasPrefix(s, fmt.Sprintf("%s-%s", d.Name, d.Variant))
+		},
+	}
+
+	CentOS = Platform{
+		Name: "centos", Id: "679593333241", DistroName: "centos", LatestVariant: "7",
+		MatchFunc: func(s string, d Distro) bool {
+			return strings.HasPrefix(s, strings.ToLower(fmt.Sprintf("%s Linux %s", d.Name, d.Variant)))
 		},
 	}